+++ foo.txt
@@ -1,0 +1,18 @@
+
+
+import <module 'dill' from '/Users/dimberman/.virtualenvs/test-pip/lib/python3.7/site-packages/dill/__init__.py'>
+import sys
+
+# Check whether Airflow is available in the environment.
+# If it is, we'll want to ensure that we integrate any macros that are being provided
+# by plugins prior to unpickling the task context.
+if sys.version_info.major >= 3 and sys.version_info.minor >= 6:
+    try:
+        from airflow.plugins_manager import integrate_macros_plugins
+        integrate_macros_plugins()
+    except ImportError:
+        # Airflow is not available in this environment, therefore we won't
+        # be able to integrate any plugin macros.
+        pass
+
+# Read args
